# openapi.yaml
openapi: 3.0.0
info:
  title: Serpent Showdown API
  version: 1.0.0
  description: API for the Serpent Showdown game, including auth, leaderboard, and live player features.
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
  /auth/login:
    post:
      summary: Login user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthCredentials'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /auth/signup:
    post:
      summary: Register new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthCredentials'
      responses:
        '200':
          description: Successful signup
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
        '400':
          description: Registration failed (e.g. email exists)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, username, email, createdAt]
    AuthCredentials:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        username:
          type: string
      required: [email, password]
...
    ApiResponseUser:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/User'
        error:
          type: string
...
    ApiResponseError:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string